Online News Popularity Analysis in R
As a part of Automated Learning and Data Analysis in R, we have
implemented SVM, ANN and Naive Bayes to predict the popularity of the
news article using the dataset from UCI's Machine Learning repository.
Using KL measure, we have tried to estimate the most important features
in the dataset. We have also tried to suggest changes in the news
article to make it popular using KNN.